STL-2731 | Optimize failure pages by creating common component
